# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Procurement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRfqs
      tags: [rfqs]
      summary: List all RFQs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of RFQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfq"
    post:
      operationId: createRfq
      tags: [rfqs]
      summary: Create a new RFQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqInput"
      responses:
        "201":
          description: RFQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{id}:
    get:
      operationId: getRfq
      tags: [rfqs]
      summary: Get RFQ by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RFQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
        "404":
          description: Not found
    patch:
      operationId: updateRfq
      tags: [rfqs]
      summary: Update RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqUpdate"
      responses:
        "200":
          description: Updated RFQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
    delete:
      operationId: deleteRfq
      tags: [rfqs]
      summary: Delete RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /rfqs/{id}/quotations/compare:
    get:
      operationId: compareQuotations
      tags: [rfqs]
      summary: Compare quotations for an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comparison data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationComparison"

  # Quotations
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List all quotations
      parameters:
        - name: rfqId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Submit a quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Quotation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get quotation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
        "404":
          description: Not found
    patch:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationUpdate"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"

  # Approvals
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List all purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      summary: Get purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      summary: Update purchase order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # Invoices
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
...
    RfqItem:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
      required:
        - name
        - quantity
        - unit
    Rfq:
      type: object
      properties:
        id:
          type: integer
        rfqNumber:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/RfqItem"
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, open, closed, awarded, cancelled]
        assignedVendorIds:
          type: array
          items:
            type: integer
        createdById:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rfqNumber
        - title
        - items
        - deadline
        - status
        - assignedVendorIds
        - createdById
        - createdAt
    RfqInput:
      type: object
      required: [title, items, deadline, assignedVendorIds]
      properties:
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RfqItem"
        deadline:
          type: string
          format: date-time
        assignedVendorIds:
          type: array
          items:
            type: integer
    RfqUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RfqItem"
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, open, closed, awarded, cancelled]
        assignedVendorIds:
          type: array
          items:
            type: integer

    # Quotation schemas
    QuotationItem:
      type: object
      properties:
        rfqItemName:
          type: string
        unitPrice:
          type: number
        quantity:
          type: number
        totalPrice:
          type: number
      required:
        - rfqItemName
        - unitPrice
        - quantity
        - totalPrice
    Quotation:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        totalAmount:
          type: number
        deliveryTimeline:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, under_review, accepted, rejected]
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - rfqId
        - vendorId
        - vendorName
        - items
        - totalAmount
        - deliveryTimeline
        - status
        - submittedAt
    QuotationInput:
      type: object
      required: [rfqId, vendorId, items, totalAmount, deliveryTimeline]
      properties:
        rfqId:
          type: integer
        vendorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        totalAmount:
          type: number
        deliveryTimeline:
          type: string
        notes:
          type: string
    QuotationUpdate:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        totalAmount:
          type: number
        deliveryTimeline:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [submitted, under_review, accepted, rejected]
    QuotationComparison:
      type: object
      properties:
        rfqId:
          type: integer
        rfqTitle:
          type: string
        quotations:
          type: array
          items:
            $ref: "#/components/schemas/Quotation"
        lowestPriceVendorId:
          type: ["integer", "null"]
        fastestDeliveryVendorId:
          type: ["integer", "null"]
      required:
        - rfqId
        - rfqTitle
        - quotations

    # Approval schemas
...
    PurchaseOrderItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        totalPrice:
          type: number
      required:
        - name
        - quantity
        - unit
        - unitPrice
        - totalPrice
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        poNumber:
          type: string
        rfqId:
          type: ["integer", "null"]
        quotationId:
          type: ["integer", "null"]
        vendorId:
          type: integer
        vendorName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        status:
          type: string
          enum: [draft, sent, acknowledged, delivered, completed, cancelled]
        deliveryDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - poNumber
        - vendorId
        - vendorName
        - items
        - subtotal
        - taxRate
        - taxAmount
        - totalAmount
        - status
        - createdAt
    PurchaseOrderInput:
      type: object
      required: [vendorId, items, taxRate]
      properties:
        rfqId:
          type: integer
        quotationId:
          type: integer
        vendorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        taxRate:
          type: number
        deliveryDate:
          type: string
          format: date-time
    PurchaseOrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, acknowledged, delivered, completed, cancelled]
        deliveryDate:
          type: string
          format: date-time

    # Invoice schemas
...